Fix "Permission denied" error when executing Reproduce as non-root user
When creating locks from a shared directory such as `/tmp/reproduce/locks`,
a `permission denied` error may occur if first one user creates that
directory and another tries the same thing. The second user will see
the error (except if Reproduce is running as root, of course).

Since the idea behind using that directory is to remove old locks on
FreeBSD startup (`clear_tmp_enable=YES`), I think it is a good idea
to use `~/.reproduce/locks` for non-root users and `/tmp/reproduce/locks`
for root.